:truck: api: move ShareableCertificate model from lib to src
