+++ api/src/certification/results/domain/models/ShareableCertificate.js
